# examples/nestjs-sample/swagger-docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/pets": {
      "get": {
        "operationId": "PetController_findAll",
        "summary": "펫 목록 조회",
        "description": "등록된 모든 펫의 목록을 조회합니다.",
        "parameters": [
          {
            "name": "species",
            "required": false,
            "in": "query",
            "description": "종류로 필터링",
            "schema": {
              "enum": [
                "dog",
                "cat",
                "bird",
                "fish"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "최대 결과 수",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "펫 목록",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "operationId": "PetController_create",
        "summary": "펫 등록",
        "description": "새로운 펫을 등록합니다.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 펫 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "잘못된 요청"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/pets/{id}": {
      "get": {
        "operationId": "PetController_findOne",
        "summary": "펫 상세 조회",
        "description": "특정 펫의 상세 정보를 조회합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "펫 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "펫 상세 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "펫을 찾을 수 없음"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "operationId": "PetController_update",
        "summary": "펫 정보 수정",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "펫 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PetResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "operationId": "PetController_remove",
        "summary": "펫 삭제",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "펫 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "삭제 완료"
          },
          "404": {
            "description": "펫을 찾을 수 없음"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/pets/{id}/avatar": {
      "post": {
        "operationId": "PetController_uploadAvatar",
        "summary": "펫 아바타 업로드",
        "description": "펫의 프로필 이미지를 업로드합니다.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "펫 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "이미지 파일"
                  },
                  "description": {
                    "type": "string",
                    "description": "이미지 설명"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "업로드 완료",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "url": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/register": {
      "post": {
        "operationId": "UserController_register",
        "summary": "회원가입",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 사용자 정보",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "이미 존재하는 이메일"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/login": {
      "post": {
        "operationId": "UserController_login",
        "summary": "로그인 (Form URL-Encoded)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT 토큰",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "인증 실패"
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/register-multi": {
      "post": {
        "operationId": "UserController_registerMulti",
        "summary": "회원가입 (JSON + Form)",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 사용자 정보 (multiple content types)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "summary": "내 정보 조회",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "summary": "사용자 조회",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "사용자 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "사용자를 찾을 수 없음"
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/petstore.v1.PetService/GetPet": {
      "post": {
        "operationId": "GrpcController_httpGetPet",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/petstore.v1.PetService/AddPet": {
      "post": {
        "operationId": "GrpcController_httpAddPet",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/petstore.v1.PetService/ListPets": {
      "post": {
        "operationId": "GrpcController_httpListPets",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/petstore.v1.StoreService/PlaceOrder": {
      "post": {
        "operationId": "GrpcController_httpPlaceOrder",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/admin/users/search": {
      "post": {
        "operationId": "AdminController_searchUsers",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/admin/users/{id}/token": {
      "post": {
        "operationId": "AdminController_generateToken",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/internal/health": {
      "get": {
        "operationId": "InternalController_healthCheck",
        "summary": "Health check",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Service is healthy"
          }
        },
        "tags": [
          "Internal"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/internal/version": {
      "get": {
        "operationId": "InternalController_getVersion",
        "summary": "Get API version",
        "parameters": [],
        "responses": {
          "200": {
            "description": "API version info"
          }
        },
        "tags": [
          "Internal"
        ],
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Petstore API",
    "description": "A sample pet store API for APIForge testing",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://api.petstore.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PetResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "펫 ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "펫 이름"
          },
          "species": {
            "type": "string",
            "description": "펫 종류"
          },
          "age": {
            "type": "number",
            "description": "펫 나이"
          },
          "createdAt": {
            "type": "string",
            "description": "등록일"
          }
        },
        "required": [
          "id",
          "name",
          "species",
          "createdAt"
        ]
      },
      "PetListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PetResponseDto"
            }
          },
          "total": {
            "type": "number",
            "description": "전체 개수"
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "CreatePetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "펫 이름",
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "description": "펫 종류",
            "enum": [
              "dog",
              "cat",
              "bird",
              "fish"
            ]
          },
          "age": {
            "type": "number",
            "description": "펫 나이",
            "example": 3
          },
          "tags": {
            "description": "태그 목록",
            "example": [
              "friendly",
              "trained"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name",
          "species"
        ]
      },
      "UpdatePetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "펫 이름"
          },
          "species": {
            "type": "string",
            "description": "펫 종류",
            "enum": [
              "dog",
              "cat",
              "bird",
              "fish"
            ]
          },
          "age": {
            "type": "number",
            "description": "펫 나이"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "minLength": 8
          },
          "displayName": {
            "type": "string",
            "description": "사용자 이름",
            "example": "John Doe"
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "사용자 ID"
          },
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "displayName": {
            "type": "string",
            "description": "사용자 이름"
          },
          "avatarUrl": {
            "type": "string",
            "description": "프로필 이미지 URL"
          },
          "createdAt": {
            "type": "string",
            "description": "가입일"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "createdAt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "password": {
            "type": "string",
            "description": "비밀번호"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT 액세스 토큰"
          },
          "expiresIn": {
            "type": "number",
            "description": "토큰 만료 시간 (초)"
          }
        },
        "required": [
          "accessToken",
          "expiresIn"
        ]
      }
    }
  }
}